refactor(prompt): extract extra prompt
